Allow specifying PG column of composite types
This allows inserting and updating composite PostgreSQL types
with an explicit naming of target columns. The code will
quote each component of a "pgcol.composite.field"

Include test for json_props_to_pg_cols and also test
for update_features to use the new syntax.

Closes #31

Drop support for ROW() literals in json_props_to_pg_cols as it
is not needed anymore.

Updates documentation for the function.